Porting to Plasma 6
Porting to Plasma 6

This commit ports the notmart.ksysguard.digitalgauge plasmoid to ensure compatibility with Plasma 6. The necessary changes have been made to adapt to the new conventions of Plasma 6 and ensure the proper functioning of the plasmoid on this version of the Plasma desktop.

Note: This commit marks the first step towards the full update of the plasmoid for Plasma 6.